% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_longiNGS.R
\name{corr_longiNGS}
\alias{corr_longiNGS}
\title{Within subject correlations for selected taxon}
\usage{
corr_longiNGS(taxa.dat, metadat, sample.id, subj.id, time.var, taxon.name)
}
\arguments{
\item{taxa.dat}{numeric matrix with taxa as rows and samples as columns
annotated with both column (sample ids) and row names (taxa names)}

\item{metadat}{A data.frame; has to contain subject ID variable, time
(visit) variable and sample ID variable}

\item{sample.id}{character string - name of sample ID variable in the metadat}

\item{subj.id}{character string - name of subject ID variable in the metadat}

\item{time.var}{character string - name of time/visit ID variable}

\item{taxon.name}{character string - name of the taxon of interest}
}
\value{
matrix of Spearman correlation estimates
}
\description{
The function outputs matrix of Spearman correlations for bacterial abundances
between timepoints for a selected taxon calculated from all available pairwise
complete observations
}
\examples{
data(ngs_dat)
counts.dat <- ngs_dat$counts
metadat <- ngs_dat$metadat
corr_longiNGS(counts.dat, metadat,
              "ID_sample", "ID_Number", "Day_Number",
              "g_Bacteroides")[1:6, 1:6]
}
