% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rec_subj.R
\name{count_rec_subj}
\alias{count_rec_subj}
\title{Count numbers of observations in the data}
\usage{
count_rec_subj(dat, ID.var, visit.var)
}
\arguments{
\item{dat}{A data.frame; has to contain subject ID variable and time
(visit) variable}

\item{ID.var}{character string - name of subjectID variable in the dat}

\item{visit.var}{character string - name of visit number/descriptor variable}
}
\value{
prints out number of records, subjects, and max number of records
  per subject
}
\description{
Count numbers of visits, subjects, and max available visits per subject.
Checks for duplicated records
}
\examples{
data(ngs_dat)
metadat <- ngs_dat$metadat
count_rec_subj(metadat, "ID_Number", "Day_Number")
}
