% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngs_dat}
\alias{ngs_dat}
\title{Longitudinal 16S Microbiome Dataset}
\format{
A list with 2 objects
\describe{
  \item{metadata}{data.frame 342 observations (samples) and 36 variables}
  \item{counts}{numeric matrix with 192 rows corresponding to taxa, and 342 columns
  corresponding to samples}
}
}
\usage{
ngs_dat
}
\description{
Data object (list of 2 datasets) containing results of 16S sequencing of
longitudinal gut microbiome profiles for 20 women over 6 weeks
}
\references{
*Vandeputte, D., De Commer, L., Tito, R.Y. et al. Temporal
variability in quantitative human gut microbiome profiles and implications
for clinical research. Nat Commun 12, 6740 (2021)*
}
\keyword{datasets}
