% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visits_per_subj.R
\name{visits_per_subj}
\alias{visits_per_subj}
\title{Count numbers of records per subject}
\usage{
visits_per_subj(dat, ID.var, visit.var)
}
\arguments{
\item{dat}{A data.frame; has to contain subject ID variable and time
(visit) variable}

\item{ID.var}{character string - name of subjectID variable in the dat}

\item{visit.var}{character string - name of visit number/descriptor variable}
}
\value{
data.frame with 2 columns
  - n_TimePoints
  - N_subjects
}
\description{
Outputs description of longitudinal observations structure:
numbers of subjects with different numbers of records. Checks for duplicated
records
}
\examples{
data(ngs_dat)
metadat <- ngs_dat$metadat
visits_per_subj(metadat, "ID_Number", "Day_Number")
}
